+++ myturkishiid/Views/FeatureViews.py
@@ -1,4 +1,5 @@
 from django.contrib import messages
+from django.contrib.auth import logout
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.shortcuts import render, redirect
@@ -13,10 +14,16 @@ from myturkishiid.Forms.FeatureTypeForm import FeatureTypeForm
 from myturkishiid.models import FeatureType, Feature, FeatureDesc, FeatureTypeDesc, Category, CategoryDesc, Advert
 
 from myturkishiid.models.Language import Language
+from myturkishiid.services import general_methods
 
 
 @login_required
 def feature_save(request):
+    perm = general_methods.control_access(request)
+
+    if not perm:
+        logout(request)
+        return redirect('accounts:login')
     form_feature = FeatureForm(request.POST)
     features = Feature.objects.all()
 
@@ -38,6 +45,11 @@ def feature_save(request):
 
 @login_required
 def featureType_save(request):
+    perm = general_methods.control_access(request)
+
+    if not perm:
+        logout(request)
+        return redirect('accounts:login')
     form_featureType = FeatureTypeForm(request.POST)
 
     if request.method == 'POST':
@@ -66,6 +78,11 @@ def get_feature(request):
 
 @login_required
 def featureDesc_save(request, pk):
+    perm = general_methods.control_access(request)
+
+    if not perm:
+        logout(request)
+        return redirect('accounts:login')
     form_featureDesc = FeatureDescForm(request.POST)
     feature = Feature.objects.get(pk=pk)
     featureDesc = FeatureDesc.objects.filter(feature=feature)
@@ -92,6 +109,11 @@ def featureDesc_save(request, pk):
 
 @login_required
 def feature_type(request):
+    perm = general_methods.control_access(request)
+
+    if not perm:
+        logout(request)
+        return redirect('accounts:login')
     feature_types = FeatureType.objects.all()
 
     form_featureType = FeatureTypeForm(request.POST)
@@ -116,6 +138,11 @@ def feature_type(request):
 
 @login_required
 def featureTypeDesc_save(request, pk):
+    perm = general_methods.control_access(request)
+
+    if not perm:
+        logout(request)
+        return redirect('accounts:login')
     form_featureTypeDesc = FeatureTypeDescForm(request.POST)
     featureType = FeatureType.objects.get(pk=pk)
     featureTypeDesc = FeatureTypeDesc.objects.filter(featureType=featureType)
@@ -141,41 +168,12 @@ def featureTypeDesc_save(request, pk):
 
 
 @login_required
-def featureType_feature_save(request, pk, ):
-    featureForm = FeatureSaveForm(request.POST)
-    featureType = FeatureType.objects.filter(pk=pk)
-    advert = Advert.objects.get(featureType=featureType)
-
-    if request.method == 'POST':
-        featureForm.save()
-        if featureForm.is_valid():
-            featureForm = featureForm.save(commit=False)
-            for feature in featureForm.cleaned_data['feature']:
-                advert.feature.add(feature)
-            featureForm.save()
-
-            messages.success(request, 'özellik Kaydedildi.')
-
-            return redirect('myturkishid:feature-save')
-
-
-        else:
-
-            messages.warning(request, 'Alanları Kontrol Ediniz')
-
-    return render(request, 'featuretemp/add-type-to-feature.html', {'featureForm': featureForm, })
-
-
-@login_required
-def get_featureType_advert(request, pk):
-    advert = Advert.objects.get(pk=pk)
-    featureTypes = FeatureType.objects.all()
-
-    return render(request, 'featuretemp/add-type-to-feature.html', {'advert': advert, 'featureTypes': featureTypes})
-
-
-@login_required
 def add_feature_to_advert(request, advert_id, featuretype_id):
+    perm = general_methods.control_access(request)
+
+    if not perm:
+        logout(request)
+        return redirect('accounts:login')
     advert = Advert.objects.get(pk=advert_id)
     exist_features = advert.feature.filter(featuretype_id=featuretype_id)
     features = Feature.objects.filter(featureType_id=featuretype_id)
@@ -184,6 +182,11 @@ def add_feature_to_advert(request, advert_id, featuretype_id):
 
 @login_required
 def add_feature_to_feature_type(request, featuretype_id):
+    perm = general_methods.control_access(request)
+
+    if not perm:
+        logout(request)
+        return redirect('accounts:login')
     exist_features = Feature.objects.filter(featureType_id=featuretype_id)
     features = Feature.objects.filter(~Q(featureType_id=featuretype_id))
     feature_type = FeatureType.objects.get(id=featuretype_id)
@@ -204,6 +207,11 @@ def add_feature_to_feature_type(request, featuretype_id):
 
 @login_required
 def delete_feature_from_feature_type(request, feature_id):
+    perm = general_methods.control_access(request)
+
+    if not perm:
+        logout(request)
+        return redirect('accounts:login')
     feature = Feature.objects.get(pk=feature_id)
     x = feature.featureType
     feature.featureType = None
@@ -214,6 +222,11 @@ def delete_feature_from_feature_type(request, feature_id):
 
 @login_required
 def featureType_desc_update(request, pk):
+    perm = general_methods.control_access(request)
+
+    if not perm:
+        logout(request)
+        return redirect('accounts:login')
     featureType_desc = FeatureTypeDesc.objects.get(pk=pk)
     featureType_desc_form = FeatureTypeDescForm(request.POST or None, instance=featureType_desc)
 
@@ -234,9 +247,13 @@ def featureType_desc_update(request, pk):
                   {'form': featureType_desc_form})
 
 
-
 @login_required
 def feature_desc_update(request, pk):
+    perm = general_methods.control_access(request)
+
+    if not perm:
+        logout(request)
+        return redirect('accounts:login')
     feature_desc = FeatureDesc.objects.get(pk=pk)
     feature_desc_form = FeatureDescForm(request.POST or None, instance=feature_desc)
 
@@ -256,9 +273,14 @@ def feature_desc_update(request, pk):
     return render(request, 'featuretemp/FeatureDesc-save.html',
                   {'form_featureDesc': feature_desc_form})
 
-def feature_delete(request,feature_id):
-        feature = Feature.objects.get(pk=feature_id)
-        feature.delete()
-        messages.success(request, 'Özellik Silindi')
-        return redirect('myturkishid:feature-save')
 
+def feature_delete(request, feature_id):
+    perm = general_methods.control_access(request)
+
+    if not perm:
+        logout(request)
+        return redirect('accounts:login')
+    feature = Feature.objects.get(pk=feature_id)
+    feature.delete()
+    messages.success(request, 'Özellik Silindi')
+    return redirect('myturkishid:feature-save')
